Fixes compatibility with the old WMI
In the case of association classes, there were some differences
between PyMI and the old WMI in the way they were returning
the association objects. This patch fixes the compatibility
with the old WMI by expecting that the association classes
return the objects and not only the references as strings.