Changes how association classes are treated
PyMI was updated in the case of association classes.
Currently when an association class is retrieved, its
properties which are references will be returned as strings.
The new PyMI update converts automatically the references
and returns the objects.

This patch changes the way in which class properties that
are references are treated, objects are expected now
instead of strings.

Change-Id: Ibb49b99604492ac72dd46482bb1f9daf98171c32